+++ odoo/local-src/connector_qoqa/__init__.py
@@ -32,3 +32,5 @@ from . import sale
 from . import sale_line
 from . import unit
 from . import wizard
+
+from . import controllers